zbt_bot/requirements.txt
aiogram~=2.20
psycopg2-binary~=2.9.3
python-dotenv~=0.20.0
requests~=2.27.1
